pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sipi-core"
version = "0.1.0"
description = "Core domain models and logic for SIPI"
dependencies = [
    "sqlalchemy[asyncio]~=2.0.36",
    "GeoAlchemy2~=0.15.2",
    "pydantic~=2.9.2",
    "python-dotenv~=1.0.1",
    "strawberry-graphql[asgi]~=0.243.0",
    "streamlit~=1.29.0"
]
requires-python = ">=3.10"

[tool.hatch.build.targets.wheel]
packages = ["src/sipi"]
